added sources with documented modifications applied
these include changing the directory of the data, as well as
some theano parameters for running on GPU, but should be
system-changes rather than model-changes (aside from the
parameter pretraining_epochs, which various values were
tried for.

See the repository
https://github.com/AIReproducibility2018/DLClassifHypspec
for specific details (as well as package versions and other
information potentially useful for getting this code the same way).

The original repository published by the authors of the paper should
still be available at:
https://github.com/hantek/deeplearn_hsi
although that repository is not frozen and has seen modifications
since the paper was published (e.g. a value that was set to
either 300 or 3300 - pretraining_epochs).

An additional issue is that the code appears to be Python2, which was
dated back when the reproduction attempt was made, and now even moreso.

I have not tried running this code again now in 2022.